Rename to whiskeyWrite
